refactor(CoinInputView): 입력 관련 클래스 세분화
